proof_of_concepts/langchain/requirements.txt
#  AI dependencies
langchain
langchain_experimental
langchain-community
openai
libmagic
unstructured
tiktoken

# Google Cloud
google-api-python-client
google-cloud-secret-manager
